Fix possible bug in GetPseudoRand().
It was possible that a deleted factoid could result in a state where
all the still-existing values for a factoid key had been 'seen' and
thus Count() would return 0, causing no factoid to be found. To fix
this, check whether the key has been 'seen' before and invalidate
that saved state if the result count is zero.